+++ data_extractor/src/main/java/com/garuda/dataextractor/extactor/DataDownloader.java
@@ -1,4 +1,4 @@
-package extractor;
+package com.garuda.dataextractor.extactor;
 
 import java.io.File;
 import java.io.FileOutputStream;
