+++ data_extractor/src/main/java/com/garuda/dataextractor/extactor/FileTransformer.java
@@ -1,4 +1,4 @@
-package extractor;
+package com.garuda.dataextractor.extactor;
 
 import java.io.IOException;
 import java.nio.file.Files;
